c-kzg ffi bindings
point to curve
Bit reversal tests
